requirements.txt
asgiref==3.4.1
Django==3.2.9
django-3-jet==1.0.8
django-admin-interface==0.18.2
django-colorfield==0.4.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.12.4
gunicorn==20.1.0
passlib==1.7.4
Pillow==8.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
